requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-telegram-bot==21.3
ccxt==4.3.0
python-dotenv==1.0.1
pydantic==2.7.4
numpy==1.26.4
